Pass URIs through the Cordova resource API.
This causes it to delegate handling of "special" URI schemes to the
relevant plugins.  This fixes webintents to work properly (at least on
Android) when using the latest version of the File plugin, which now
uses a custom cdvfile:// scheme to automatically and transparently
choose the best location to store files.